+++ utils/UnitsUtil.cpp
@@ -1,16 +1,16 @@
 #include "UnitsUtil.h"
 #include <math.h>
 
-float UnitsUtil::MILLI = 1e-3f;
-float UnitsUtil::MICRO = 1e-6f;
-float UnitsUtil::NANO = 1e-9f;
-float UnitsUtil::PICO = 1e-12f;
+double UnitsUtil::MILLI = 1e-3f;
+double UnitsUtil::MICRO = 1e-6f;
+double UnitsUtil::NANO = 1e-9f;
+double UnitsUtil::PICO = 1e-12f;
 
 UnitsUtil::UnitsUtil()
 {
 }
 
-QString UnitsUtil::toString(float f, const QString& unit) {
+QString UnitsUtil::toString(double f, const QString& unit) {
   if(fabs(f) < PICO) {
     return QString("0") + unit;
   }
@@ -29,7 +29,7 @@ QString UnitsUtil::toString(float f, const QString& unit) {
   return QString::number(f, 'g', 3) + unit;
 }
 
-float UnitsUtil::roundToOrderOfMagnitude(float f) {
+double UnitsUtil::roundToOrderOfMagnitude(double f) {
   if(f >= 1) {
     return pow(10, ceil(log10(f)));
   } else {
